changed threshold distance type to float from double in Line2DData::getPos
git-svn-id: https://svn.opendtect.org/od/trunk@27702 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiMPE/horflatvieweditor2d.cc
@@ -583,7 +583,7 @@ bool HorizonFlatViewEditor2D::getPosID( const Coord3& crd,
     if ( !linesetgeom.getLineData( linenm_ ) )
 	return false;
 
-    linesetgeom.getLineData( linenm_ )->getPos( crd, pos, mUdf(double) );
+    linesetgeom.getLineData( linenm_ )->getPos( crd, pos, mUdf(float) );
     mDynamicCastGet(const EM::Horizon2D*,hor2d,emobj);
 
     if ( !hor2d ) return false;
